[RPSDK-1298]: [ZEPHYR] Add support for Zephyr External Module
This commit adds support for this repo rpmsg-lite
to be importable in Zephyr as External Module.

Add zephyr/module.yml and Kconfig for Zephyr to recognize this repo.
This will make an entry in Zephyr menuconfig for RPMSG-LITE Config.

Add zephyr/samples with RPSMG-Lite Ping Pong Example to test library
build and run.

Signed-off-by: Tomas Galbicka <tomas.galbicka@nxp.com>